backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
MetaTrader5==5.0.4424
pymongo==4.7.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.3
python-dotenv==1.0.1
requests==2.31.0
python-multipart==0.0.9
